Updated package-manager to support cocoapods
