--- packages/expo-pod-install/README.md
@@ -1,12 +1,0 @@
-<!-- Title -->
-<h1 align="center">
-👋 Welcome to <br><code>@expo/pod-install</code>
-</h1>
-
----
-
-Ensure CocoaPods are installed in your project.
-
-## License
-
-The Expo source code is made available under the [MIT license](LICENSE). Some of the dependencies are licensed differently, with the BSD license, for example.
