--- packages/expo-pod-install/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "@expo/pod-install",
-  "version": "0.0.0",
-  "main": "build/CocoaPods.js",
-  "keywords": [
-    "cocoapods",
-    "pod-install",
-    "ios"
-  ],
-  "description": "Ensure CocoaPods are installed in your project",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/expo/expo-cli.git"
-  },
-  "author": "Expo <support@expo.io>",
-  "license": "MIT",
-  "files": [
-    "build"
-  ],
-  "scripts": {
-    "lint": "eslint .",
-    "watch": "tsc --watch",
-    "build": "tsc",
-    "prepare": "yarn run clean && yarn build",
-    "clean": "rm -rf build ./tsconfig.tsbuildinfo"
-  },
-  "devDependencies": {
-    "@expo/babel-preset-cli": "0.2.7",
-    "rimraf": "^3.0.2"
-  },
-  "dependencies": {
-    "@expo/spawn-async": "1.5.0",
-    "chalk": "2.4.2",
-    "sudo-prompt": "9.1.1",
-    "update-check": "1.5.3"
-  }
-}
