+++ packages/package-manager/package.json
@@ -42,7 +42,8 @@
     "fs-extra": "^8.1.0",
     "npm-package-arg": "^7.0.0",
     "split": "^1.0.1",
-    "stream": "^0.0.2"
+    "stream": "^0.0.2",
+    "sudo-prompt": "9.1.1"
   },
   "devDependencies": {
     "@expo/babel-preset-cli": "0.2.7"
