+++ packages/package-manager/src/index.ts
@@ -1,0 +1,3 @@
+export * from './PackageManager';
+export * from './NodePackageManagers';
+export * from './CocoaPodsPackageManager';
